+++ src/main/java/me/rkfg/xmpp/bot/plugins/AnonEchoPlugin.java
@@ -7,7 +7,6 @@ import java.util.regex.Matcher;
 import org.jivesoftware.smack.packet.Message;
 import org.jxmpp.util.XmppStringUtils;
 
-import me.rkfg.xmpp.bot.Main;
 import ru.ppsrk.gwt.client.ClientAuthException;
 import ru.ppsrk.gwt.client.LogicException;
 
@@ -15,7 +14,7 @@ public class AnonEchoPlugin extends CommandPlugin {
     @Override
     public String processCommand(Message message, Matcher matcher) throws LogicException, ClientAuthException {
         String s = matcher.group(COMMAND_GROUP);
-        Main.sendMUCMessage(s, XmppStringUtils.parseBareJid(message.getFrom()));
+        sendMUCMessage(s, XmppStringUtils.parseBareJid(message.getFrom()));
         return "Ok";
     }
