+++ src/main/java/me/rkfg/xmpp/bot/plugins/GoodbyePlugin.java
@@ -3,8 +3,6 @@ package me.rkfg.xmpp.bot.plugins;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import me.rkfg.xmpp.bot.Main;
-
 import org.jivesoftware.smack.packet.Message;
 
 public class GoodbyePlugin extends MessagePluginImpl {
@@ -22,12 +20,12 @@ public class GoodbyePlugin extends MessagePluginImpl {
     @Override
     public void init() {
         super.init();
-        Main.getSettingsManager().setDefault("goodbye_msg", "Я всё.");
+        getSettingsManager().setDefault("goodbye_msg", "Я всё.");
         Runtime.getRuntime().addShutdownHook(new Thread(new Runnable() {
 
             @Override
             public void run() {
-                Main.sendMUCMessage(Main.getSettingsManager().getStringSetting("goodbye_msg"));
+                sendMUCMessage(getSettingsManager().getStringSetting("goodbye_msg"));
                 try {
                     Thread.sleep(3000);
                 } catch (InterruptedException e) {
