+++ src/main/java/me/rkfg/xmpp/bot/plugins/MessagePluginImpl.java
@@ -1,10 +1,16 @@
 package me.rkfg.xmpp.bot.plugins;
 
+import org.jivesoftware.smack.chat.ChatManager;
 import org.jivesoftware.smack.packet.Message;
 import org.jivesoftware.smack.packet.Message.Type;
 import org.jxmpp.util.XmppStringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import me.rkfg.xmpp.bot.Bot;
+import me.rkfg.xmpp.bot.ChatAdapter;
+import me.rkfg.xmpp.bot.MUCManager;
+import ru.ppsrk.gwt.server.SettingsManager;
 
 public abstract class MessagePluginImpl implements MessagePlugin {
     protected Logger log = LoggerFactory.getLogger(getClass());
@@ -56,5 +62,33 @@ public abstract class MessagePluginImpl implements MessagePlugin {
             return nick;
         }
         return nick.substring(0, 1) + "\u200b" + nick.substring(1);
+    }
+
+    protected SettingsManager getSettingsManager() {
+        return Bot.INSTANCE.getSettingsManager();
+    }
+
+    protected MUCManager getMUCManager() {
+        return Bot.INSTANCE.getMUCManager();
+    }
+
+    protected ChatManager getChatManagerInstance() {
+        return Bot.INSTANCE.getChatManagerInstance();
+    }
+
+    protected void sendMUCMessage(String message, String mucName) {
+        Bot.INSTANCE.sendMUCMessage(message, mucName);
+    }
+
+    protected void sendMessage(ChatAdapter chatAdapter, String message) {
+        Bot.INSTANCE.sendMessage(chatAdapter, message);
+    }
+
+    protected void sendMUCMessage(String message) {
+        Bot.INSTANCE.sendMUCMessage(message);
+    }
+
+    protected String getBotNick() {
+        return Bot.INSTANCE.getBotNick();
     }
 }
